Update projectile search to index quickly
While this does mean that it may miss some files here and there,
it will be incredibly fast, which is what I would prefer. For
other files, you will need to use C-x C-f and just need to know
where they are, as this is an unfortunate circumstance brought
about between Emacs, Projectile, and the OS for searching for
files. You could instead use native searching and enable caching,
but I chose to avoid that as the cache is liable to grow stale.